Attempting to split arm baremetal PAL
